Project/PythonProject/ihub/requirements.txt
antiorm==1.2.1
asgiref==3.2.9
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
demjson==2.2.4
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-bootstrap4==2.1.1
django-heroku==0.3.1
django-imagekit==4.0.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
install==1.3.3
isort==4.3.21
jtutils==0.0.6
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
leven==1.0.4
matplotlib==3.2.2
mccabe==0.6.1
nose==1.3.7
numpy==1.19.0
pandas==1.0.5
pilkit==2.0
Pillow==7.1.2
psycopg2==2.7.5
pylint==2.5.3
pyparsing==2.4.7
python-csv==0.0.11
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
webdriver-manager==3.2.1
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
zipp==3.1.0
